backend/requirements.txt
fastapi==0.128.8
uvicorn==0.39.0
openai==2.23.0
chromadb==1.5.1
PyGithub==2.8.1
python-dotenv==1.2.1
vulture==2.14
